+++ clib/v2_motion_done.c
@@ -18,7 +18,7 @@ void set_chekr_v2_motion_done__()
 }
 
 int v2_motion_done(ip,indx)
-long ip[5];                          /* ipc array */
+int ip[5];                          /* ipc array */
 int indx;
 {
       struct req_buf buffer;
@@ -27,7 +27,7 @@ int indx;
       struct res_rec response;
       int counts,motion, oorange, time_out;
       struct tms tms_buff;
-      long end;
+      int end;
 
       ini_req(&buffer);                      /* format the buffer */
       if(indx == 0)
